Makefile more pedantic

+++ problem3.c
@@ -10,8 +10,8 @@ struct box {
 #define DUMB_LOC(x,y) (((y) << 10) + x)
 
 int check_box(struct box *b, short *data) {
-	for (int i = 0; i < b->x_len; ++i) {
-		for (int j = 0; j < b->y_len; ++j) {
+	for (size_t i = 0; i < b->x_len; ++i) {
+		for (size_t j = 0; j < b->y_len; ++j) {
 			if (data[DUMB_LOC(b->x_pos + i, b->y_pos + j)] != 1)
 				return 0;
 		}
@@ -24,20 +24,19 @@ void solve_problem3(FILE *input_file) {
 	struct box tmp_box;
 	vector_init(&boxes, sizeof(struct box), 10);
 	while (fscanf(input_file, "#%d @ %d,%d: %dx%d\n", &tmp_box.id, &tmp_box.x_pos, &tmp_box.y_pos, &tmp_box.x_len, &tmp_box.y_len) == 5) {
-		//fprintf("Scanned box %d: [%d,%d) -> [%d,%d)\n", tmp_box.id, tmp_box.y_pos, tmp_box.y_len, tmp_box.x_pos, tmp_box.x_len);
 		vector_push(&boxes, &tmp_box);
 	}
 	short data[1024*1024] = {0};
 	VECTOR_FOREACH(struct box *, b, boxes) {
 		//fprintf(stderr, "box: %d\n", b->id);
-		for (int i = 0; i < b->x_len; ++i) {
-			for (int j = 0; j < b->y_len; ++j) {
+		for (size_t i = 0; i < b->x_len; ++i) {
+			for (size_t j = 0; j < b->y_len; ++j) {
 				data[DUMB_LOC(b->x_pos + i, b->y_pos + j)]++;
 			}
 		}
 	}
 	int total = 0;
-	for(int i = 0; i<1024 * 1024; ++i) {
+	for(size_t i = 0; i<1024 * 1024; ++i) {
 		total += (data[i] > 1);
 	}
